Move AST-related utilities into the new ast.ts module
