+++ src/stores/CatalogOnlineQuery/CatalogOnlineQueryConfigStore.ts
@@ -3,8 +3,8 @@ import {action, computed, makeObservable, observable, reaction} from "mobx";
 
 import {CatalogSystemType, Point2D} from "models";
 import {CatalogDatabase} from "services";
-import {AppStore, ASTSettingsString, NumberFormatType, SystemType} from "stores";
-import {clamp, getPixelValueFromWCS, setAstSystem, transformPoint, VizierResource} from "utilities";
+import {AppStore, NumberFormatType, SystemType} from "stores";
+import {ASTSettingsString, clamp, getPixelValueFromWCS, setAstSystem, transformPoint, VizierResource} from "utilities";
 
 export enum RadiusUnits {
     DEGREES = "deg",
